Can use gitrev env variable to retrieve git revision
Instead of directly using the TH splice for `gitRevision`, this adds a
replacement function that first tries to retrieve that value if it
fails (eg. finds UNKNOWN) tries to retrieve the revision from an
environment variable called 'gitrev'.

The idea is that this can be used in a build environment which does
not have directly access to git executable, eg. in CI.